requirements.txt
asgiref==3.2.7
Django==3.0.5
django-widget-tweaks==1.4.8
pytz==2020.1
sqlparse==0.3.1
# Advanced Features Dependencies
djangorestframework==3.12.4
django-cors-headers==3.7.0
channels==2.4.0
channels-redis==3.2.0
celery==5.2.0
redis==3.5.3
Pillow==8.3.2
django-filter==2.4.0
django-crispy-forms==1.11.2
crispy-bootstrap4==2021.1
reportlab==3.6.8
openpyxl==3.0.9
django-ratelimit==3.0.1
django-otp==1.0.3
qrcode==7.3.1
plotly==5.6.0
pandas==1.3.4
numpy==1.21.4
# Production Dependencies
gunicorn==20.1.0
whitenoise==6.0.0
dj-database-url==0.5.0
psycopg2-binary==2.9.3
